requirements.txt
praw==7.8.1
pandas==2.2.3
matplotlib==3.9.4
python-dotenv==1.0.1
pytest==9.0.2
